Version 1.2 improvements
Switched to using Eurostat Bulkdownload Facility for International trade
instead of Easy Comext and to using OData for collecting information
about solar panels capacity. Extrapolations are completely redone, now
various new approaches for forecasting and also a new approach for
extrapolations in the past. Calculation of number of pieces for WEEE is
now based on the average weight in the year product was put on the
market instead of on the average weight of the product in the year it
became WEEE. Calculation of number of pieces for POM, WEEE and STOCK is
now also done for the aggregates of the 54 UNU_Keys. Also some changes
in average weights, code links and manual corrections.